chore: moving express and resolving path errors
