+++ src/cartridges/app_adyen_SFRA/cartridge/client/default/js/adyen/checkout/index.js
@@ -1,4 +1,4 @@
-const store = require('../../../../config/store');
+const store = require('../../../../../config/store');
 const {
   setCheckoutConfiguration,
   actionHandler,
@@ -9,10 +9,10 @@ const {
   showValidation,
   paymentFromComponent,
 } = require('./helpers');
-const billing = require('../checkout/billing');
-const { httpClient } = require('../commons/httpClient');
-const { getPaymentMethods } = require('../commons');
-const { GIFTCARD } = require('../../../../config/constants');
+const billing = require('../../checkout/billing');
+const { httpClient } = require('../../commons/httpClient');
+const { getPaymentMethods } = require('../../commons');
+const { GIFTCARD } = require('../../../../../config/constants');
 const { renderGiftCards } = require('./giftcards');
 const { addStores } = require('./pos');
