+++ src/cartridges/app_adyen_SFRA/cartridge/client/default/js/adyen/express/__tests__/amazonPayExpressPart2.test.js
