+++ src/cartridges/app_adyen_SFRA/cartridge/client/default/js/adyen/express/__tests__/applePayExpress.test.js
@@ -3,14 +3,13 @@
  */
 
 const ApplePay = require('../paymentMethods/applepay/applepay');
-const store = require('../../../../../../cartridge/config/store');
-const helpers = require('../../../js/adyen/checkout/helpers');
-
-jest.mock('../../../js/commons/httpClient');
-jest.mock('../../../../../../cartridge/config/store');
-jest.mock('../../../js/adyen/checkout/helpers');
+const store = require('../../../../../../../cartridge/config/store');
+const helpers = require('../../../../js/adyen/checkout/helpers');
+jest.mock('../../../../js/commons/httpClient');
+jest.mock('../../../../../../../cartridge/config/store');
+jest.mock('../../../../js/adyen/checkout/helpers');
 jest.mock('../initializeCheckout');
-jest.mock('../../commons');
+jest.mock('../../../commons');
 
 describe('ApplePay class', () => {
   let applePay;
