+++ src/cartridges/app_adyen_SFRA/cartridge/client/default/js/adyen/express/__tests__/googlePayExpress.test.js
@@ -3,14 +3,14 @@
  */
 
 const GooglePay = require('../paymentMethods/googlepay/googlepay');
-const store = require('../../../../../config/store');
-const helpers = require('../../../js/adyen/checkout/helpers');
+const store = require('../../../../../../config/store');
+const helpers = require('../../../../js/adyen/checkout/helpers');
 
-jest.mock('../../../js/commons/httpClient');
-jest.mock('../../../../../config/store');
-jest.mock('../../../js/adyen/checkout/helpers');
+jest.mock('../../../../js/commons/httpClient');
+jest.mock('../../../../../../config/store');
+jest.mock('../../../../js/adyen/checkout/helpers');
 jest.mock('../initializeCheckout');
-jest.mock('../../../js/commons');
+jest.mock('../../../../js/commons');
 
 
 describe('GooglePay class', () => {
