+++ src/cartridges/app_adyen_SFRA/cartridge/client/default/js/adyen/express/initializeCheckout.js
