+++ src/cartridges/app_adyen_SFRA/cartridge/client/default/js/adyen/express/paymentMethods/amazonpay/amazonPayExpressPart1.js
@@ -1,5 +1,5 @@
-const store = require('../../../../../../config/store');
-const { AMAZON_PAY } = require('../../../../../../config/constants');
+const store = require('../../../../../../../config/store');
+const { AMAZON_PAY } = require('../../../../../../../config/constants');
 const { initializeCheckout } = require('../../initializeCheckout');
 
 class AmazonPay {
