+++ src/cartridges/app_adyen_SFRA/cartridge/client/default/js/adyen/express/paymentMethods/applepay/applepay.js
@@ -1,9 +1,9 @@
-const { httpClient } = require('../../../commons/httpClient');
-const store = require('../../../../../../config/store');
-const helpers = require('../../../adyen/checkout/helpers');
-const { APPLE_PAY } = require('../../../../../../config/constants');
+const { httpClient } = require('../../../../commons/httpClient');
+const store = require('../../../../../../../config/store');
+const helpers = require('../../../checkout/helpers');
+const { APPLE_PAY } = require('../../../../../../../config/constants');
 const { initializeCheckout } = require('../../initializeCheckout');
-const { createTemporaryBasket } = require('../../../commons');
+const { createTemporaryBasket } = require('../../../../commons');
 
 class ApplePay {
   constructor(config, applicationInfo, adyenTranslations, isExpressPdp) {
