+++ src/cartridges/app_adyen_SFRA/cartridge/client/default/js/cartExpressPayments.js
@@ -3,5 +3,5 @@ const processInclude = require('base/util');
 const $ = require('jquery');
 
 $(document).ready(() => {
-  processInclude(require('./adyen_express/cart/expressPayments'));
+  processInclude(require('./adyen/express/cart/expressPayments'));
 });
