+++ src/cartridges/app_adyen_SFRA/cartridge/templates/default/cart/checkoutButtons.isml
@@ -26,7 +26,7 @@
     </script>
     <isscript>
         var assets = require('*/cartridge/scripts/assets.js');
-        assets.addJs('/js/adyen_express/paymentMethods/amazonpay/amazonPayExpressPart2.js');
+        assets.addJs('/js/adyen/express/paymentMethods/amazonpay/amazonPayExpressPart2.js');
     </isscript>
     <input id="adyenStateData" type="hidden" name="${adyenPaymentFields.adyenStateData.htmlName}"/>
     <div id="amazonPayAddressDetails">
